Updates to Dummy and telnet plugins.
Added time stamp to both plugins. Also added a formatted time stamp to the properties section of the geoIp Feature.  Also added a check to verify that the geoIP
DB returned a valid result.  If the ip is not found the feature creation process is halted and an empty string is stored in its place.